+++ examples/openapi-ts-sample/vite.config.ts
@@ -1,7 +1,7 @@
+import { createViteConfig } from '@config/vite-base';
 import react from '@vitejs/plugin-react';
-import { defineConfig } from 'vite';
 
 // https://vitejs.dev/config/
-export default defineConfig({
+export default createViteConfig({
   plugins: [react()],
 });