+++ examples/openapi-ts-tanstack-svelte-query/vite.config.ts
@@ -1,13 +1,14 @@
-import { sveltekit } from '@sveltejs/kit/vite';
-import { defineConfig } from 'vitest/config';
+import { fileURLToPath } from 'node:url';
 
-export default defineConfig({
-  // @ts-ignore
-  plugins: [
-    // @ts-expect-error
-    sveltekit(),
-  ],
-  test: {
-    include: ['src/**/*.{test,spec}.{js,ts}'],
-  },
-});
+import { createVitestConfig } from '@config/vite-base';
+import { sveltekit } from '@sveltejs/kit/vite';
+
+export default async () => {
+  const sveltekitPlugins = await sveltekit();
+  return createVitestConfig(fileURLToPath(new URL('./', import.meta.url)), {
+    plugins: [sveltekitPlugins],
+    test: {
+      include: ['src/**/*.{test,spec}.{js,ts}'],
+    },
+  });
+};