+++ packages/openapi-ts-tests/test/index.test.ts
@@ -53,10 +53,12 @@ describe('OpenAPI v2', () => {
       input: V2_SPEC_PATH,
       output,
     });
-    sync(`${output}**/*.ts`).forEach((file) => {
-      const content = fs.readFileSync(file, 'utf8').toString();
-      expect(content).toMatchFileSnapshot(toSnapshotPath(file));
-    });
+    await Promise.all(
+      sync(`${output}**/*.ts`).map(async (file) => {
+        const content = fs.readFileSync(file, 'utf8').toString();
+        await expect(content).toMatchFileSnapshot(toSnapshotPath(file));
+      }),
+    );
   });
 });
 
@@ -455,10 +457,12 @@ describe('OpenAPI v3', () => {
           path: output,
         },
       });
-      sync(`${output}**/*.ts`).forEach((file) => {
-        const content = fs.readFileSync(file, 'utf8').toString();
-        expect(content).toMatchFileSnapshot(toSnapshotPath(file));
-      });
+      await Promise.all(
+        sync(`${output}**/*.ts`).map(async (file) => {
+          const content = fs.readFileSync(file, 'utf8').toString();
+          await expect(content).toMatchFileSnapshot(toSnapshotPath(file));
+        }),
+      );
     },
   );
 
@@ -499,68 +503,80 @@ describe('OpenAPI v3', () => {
         ],
       });
 
-      sync(`${output}**/*.ts`).forEach((file) => {
-        const content = fs.readFileSync(file, 'utf8').toString();
-        expect(content).toMatchFileSnapshot(toSnapshotPath(file));
-      });
+      await Promise.all(
+        sync(`${output}**/*.ts`).map(async (file) => {
+          const content = fs.readFileSync(file, 'utf8').toString();
+          await expect(content).toMatchFileSnapshot(toSnapshotPath(file));
+        }),
+      );
     },
   );
 });
 
 describe('index', () => {
   it('parses v2 without issues', async () => {
-    await createClient({
-      dryRun: true,
-      input: './test/spec/v2.json',
-      output: './generated/v2/',
-      plugins: ['@hey-api/client-fetch'],
-    });
+    await expect(
+      createClient({
+        dryRun: true,
+        input: './test/spec/v2.json',
+        output: './generated/v2/',
+        plugins: ['@hey-api/client-fetch'],
+      }),
+    ).resolves.not.toThrow();
   });
 
   it('parses v3 without issues', async () => {
-    await createClient({
-      dryRun: true,
-      input: './test/spec/v3.json',
-      output: './generated/v3/',
-      plugins: ['@hey-api/client-fetch'],
-    });
+    await expect(
+      createClient({
+        dryRun: true,
+        input: './test/spec/v3.json',
+        output: './generated/v3/',
+        plugins: ['@hey-api/client-fetch'],
+      }),
+    ).resolves.not.toThrow();
   });
 
   it('parses v3-transforms without issues', async () => {
-    await createClient({
-      dryRun: true,
-      input: './test/spec/v3-transforms.json',
-      output: './generated/v3/',
-      plugins: [
-        '@hey-api/client-fetch',
-        '@hey-api/schemas',
-        '@hey-api/sdk',
-        '@hey-api/typescript',
-        {
-          dates: true,
-          name: '@hey-api/transformers',
-        },
-      ],
-    });
+    await expect(
+      createClient({
+        dryRun: true,
+        input: './test/spec/v3-transforms.json',
+        output: './generated/v3/',
+        plugins: [
+          '@hey-api/client-fetch',
+          '@hey-api/schemas',
+          '@hey-api/sdk',
+          '@hey-api/typescript',
+          {
+            dates: true,
+            name: '@hey-api/transformers',
+          },
+        ],
+      }),
+    ).resolves.not.toThrow();
   });
 
   it('downloads and parses v2 without issues', async () => {
-    await createClient({
-      dryRun: true,
-      input:
-        'https://raw.githubusercontent.com/hey-api/openapi-ts/main/packages/openapi-ts-tests/test/spec/v2.json',
-      output: './generated/v2-downloaded/',
-      plugins: ['@hey-api/client-fetch'],
-    });
+    await expect(
+      createClient({
+        dryRun: true,
+        input:
+          'https://raw.githubusercontent.com/hey-api/openapi-ts/main/packages/openapi-ts-tests/test/spec/v2.json',
+        output: './generated/v2-downloaded/',
+        plugins: ['@hey-api/client-fetch'],
+      }),
+    ).resolves.not.toThrow();
   });
 
   it('downloads and parses v3 without issues', async () => {
-    await createClient({
-      dryRun: true,
-      input:
-        'https://raw.githubusercontent.com/hey-api/openapi-ts/main/packages/openapi-ts-tests/test/spec/v3.json',
-      output: './generated/v3-downloaded/',
-      plugins: ['@hey-api/client-fetch'],
-    });
+    await expect(
+      createClient({
+        dryRun: true,
+        input:
+          'https://raw.githubusercontent.com/hey-api/openapi-ts/main/packages/openapi-ts-tests/test/spec/v3.json',
+        output: './generated/v3-downloaded/',
+        plugins: ['@hey-api/client-fetch'],
+      }),
+    ).resolves.not.toThrow();
   });
 });